+++ 05-ScenceManager/PiranhaPlant.cpp
@@ -13,7 +13,7 @@ PiranhaPlant::PiranhaPlant(float pipeX, float pipeY, float pipeWidth, float pipe
 	this->pipeHeight = pipeHeight;
 
 	SetState(PIRANHAPLAN_STATE_AWAKE);
-	climax = pipeY - ((level == PIRANHAPLANT_LEVEL_BIG) ? PIRANHAPLANT_BBOX_HEIGHT : PIRANHAPLANT_BBOX_HEIGHT_SMALL);
+	climax = (int)(pipeY - ((level == PIRANHAPLANT_LEVEL_BIG) ? PIRANHAPLANT_BBOX_HEIGHT : PIRANHAPLANT_BBOX_HEIGHT_SMALL));
 
 	y = pipeY;
 
@@ -135,7 +135,7 @@ void PiranhaPlant::Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects)
 	y += dy;
 
 	if (this->y <= climax && !startAttacking) {
-		y = climax;
+		y = (float)climax;
 		vy = 0;
 		startAttacking = (DWORD)GetTickCount64();
 		startResting = 0;
